updating models

+++ TradCoTracker/db/migrate/20210213182237_create_business_events.rb
@@ -1,8 +1,8 @@
 class CreateBusinessEvents < ActiveRecord::Migration[6.1]
   def change
     create_table :business_events do |t|
-      t.int :businessEventID
-      t.int :numOfPoints
+      t.numeric  :businessEventID
+      t.numeric  :numOfPoints
       t.date :submissionDate
 
       t.timestamps
