+++ TradCoTracker/db/migrate/20210213182258_create_speaking_events.rb
@@ -1,8 +1,8 @@
 class CreateSpeakingEvents < ActiveRecord::Migration[6.1]
   def change
     create_table :speaking_events do |t|
-      t.int :speakingEventID
-      t.int :numOfPoints
+      t.numeric  :speakingEventID
+      t.numeric  :numOfPoints
       t.date :submissionDate
 
       t.timestamps
